bot/requirements.txt
pyrogram==2.0.106
google-genai==1.2.0
tgcrypto-pyrofork==1.2.7
#uvloop==0.21.0
apscheduler==3.11.0
GitPython==3.1.44
pyyaml==6.0.2
pymongo==4.10.1
lupa==2.4
origamibot==2.3.6